+++ packages/utils/test/performance.spec.ts
@@ -19,12 +19,16 @@ function advanceTimersByTime(ms: number) {
 }
 
 describe('performance', () => {
+  beforeEach(() => {
+    mockFn.mockReset()
+  })
+
   it('sleep', async () => {
     async function diff(ms: number) {
       const start = Date.now()
       await sleep(ms)
       const end = Date.now()
-      return end - start > ms
+      return end - start >= ms
     }
 
     expect(await diff(0)).toBeTruthy()
@@ -33,10 +37,19 @@ describe('performance', () => {
     expect(await diff(3000)).toBeTruthy()
   }, 100000)
 
-  it('debounce', async () => {
-    const delayedFn = debounce(mockFn, 100)
+  describe('debounce', () => {
+    beforeEach(() => {
+      vi.useFakeTimers()
+    })
+
+    afterEach(() => {
+      vi.useRealTimers()
+      vi.clearAllTimers()
+    })
 
     it('should delay execution by specified time', async () => {
+      const delayedFn = debounce(mockFn, 100)
+
       delayedFn()
       expect(mockFn).not.toBeCalled()
       await advanceTimersByTime(100)
@@ -44,6 +57,8 @@ describe('performance', () => {
     })
 
     it('should execute only the last call when called continuously', async () => {
+      const delayedFn = debounce(mockFn, 100)
+
       delayedFn()
       delayedFn()
       delayedFn()
@@ -52,23 +67,26 @@ describe('performance', () => {
     })
 
     it('should pass arguments to the debounced function', async () => {
+      const delayedFn = debounce(mockFn, 100)
+
       delayedFn(1, 2, 3)
       await advanceTimersByTime(100)
       expect(mockFn).toBeCalledWith(1, 2, 3)
     })
   })
 
-  it('throttle', async () => {
-    let testFunction: Mock<any, any>
+  describe('throttle', () => {
+    let testFunction: Mock
     let throttledFunction: (this: unknown, ...args: any) => void
 
     beforeEach(() => {
+      vi.useFakeTimers()
       testFunction = vi.fn()
       throttledFunction = throttle(testFunction, 1000)
-      vi.useFakeTimers()
     })
 
     afterEach(() => {
+      vi.useRealTimers()
       vi.clearAllTimers()
     })